\name{annotatedDataFrameFrom}
\alias{annotatedDataFrameFrom}
\title{Use BigMatrix in eSet assayData}
\description{
  The Biobase package's eSet objects hold one or more
  matrices within an environment stored in their assayData
  slot. Use of classes other than a base matrix require a
  method to create an annotatedDataFrame from that class.
  Users are unlikely to ever call this method.
}

