% Generated by roxygen2 (4.0.1): do not edit by hand
\name{updateBigMatrix}
\alias{updateBigMatrix}
\title{Update previous BigMatrix type to new BigMatrix}
\usage{
updateBigMatrix(object)
}
\arguments{
  \item{object}{BigMatrix}
}
\value{
BigMatrix
}
\description{
BigMatrix has changed some of its internal storage to
eliminate the descriptor file and to keep the dimnames on
the R side. This function will take a Version <= 1.3 type
and update it to the Version >=1.4 type.
}

